Updated to allow vCheck.ps1 passed in values for Server and DataCenter
Modified Connection Plugin to allow for vCheck's new feature of having the Server and DataCenter parameters to be overridden/passed in on the command line.  Removed $Server from Connection Plugin.  If DataCenter is defined then add the -Location "$DataCenter" to the command lines of the Get-VM, Get-VMHost, Get-Cluster and Get-Datastore commands.